Added node identifications.
Nodes can be identified and accesses from the outside of the scheduler.

+++ Tests/duplicate_sync/scheduler.cpp
@@ -71,6 +71,8 @@ The support classes and code is covered by CMSIS-DSP license.
 #define CG_AFTER_NODE_EXECUTION
 #endif
 
+
+
 CG_AFTER_INCLUDES
 
 
@@ -79,10 +81,11 @@ CG_AFTER_INCLUDES
 Description of the scheduling. 
 
 */
-static unsigned int schedule[7]=
+static uint8_t schedule[7]=
 { 
 6,0,1,2,3,4,5,
 };
+
 
 CG_BEFORE_FIFO_BUFFERS
 /***********
